requirements.txt
streamlit==1.22.0
streamlit-option-menu==0.3.6
pandas==1.5.3
Pillow==9.5.0
scikit-learn==1.2.2